+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - created tag `ma27_api_key_authentication.password_hasher` which allows to create custom services and use them via the configuration
   - no BC breaks (all currently existing hashers can be used by the same config)
 
+- [bug] ensure that no api key is generated before a new one will be created (#48)
+
 ## 1.1.0
 
 - [feature] made api_key header configurable: (#32)