+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 
 - [bug] ensure that no api key is generated before a new one will be created (#48)
 
-- [feature] added `response.api_key_property` and `response.error_property` to the configuration to keep the response configurable
+- [feature] added `response.api_key_property` and `response.error_property` to the configuration to keep the response configurable (#45)
+
+- [minor] deprecated `api_key_purge.logger_service`, logger support will be removed in 2.0 (#50) 
 
 ## 1.1.0